requirements.txt
lxml==4.2.1
pdf2image==0.1.14
Pillow==5.1.0
python-pptx==0.6.10
XlsxWriter==1.0.5
